refactor: Split into modules
